+++ components/testing/src/ome/testing/ObjectFactory.java
@@ -109,7 +109,7 @@ public class ObjectFactory {
             pl.setTheC(new Integer(0));
             pl.setTheZ(new Integer(0));
             pl.setTheT(new Integer(0));
-            pl.setDeltaT(new Float(0.0));
+            pl.setDeltaT(new Double(0.0));
             p.addPlaneInfo(pl);
             i.setName("test");
             i.setAcquisitionDate(new Timestamp(System.currentTimeMillis()));
@@ -121,15 +121,15 @@ public class ObjectFactory {
         p.setSizeZ(new Integer(1));
         p.setSizeC(new Integer(1));
         p.setSizeT(new Integer(1));
-        p.setPhysicalSizeX(1.0f);
-        p.setPhysicalSizeY(1.0f);
-        p.setPhysicalSizeZ(1.0f);
+        p.setPhysicalSizeX(1.0);
+        p.setPhysicalSizeY(1.0);
+        p.setPhysicalSizeZ(1.0);
         p.setSha1("09bc7b2dcc9a510f4ab3a40c47f7a4cb77954356"); // "pixels"
         p.setPixelsType(pt);
         p.setDimensionOrder(dO);
-        p.setPhysicalSizeX(new Float(1.0));
-        p.setPhysicalSizeY(new Float(1.0));
-        p.setPhysicalSizeZ(new Float(1.0));
+        p.setPhysicalSizeX(new Double(1.0));
+        p.setPhysicalSizeY(new Double(1.0));
+        p.setPhysicalSizeZ(new Double(1.0));
         p.setImage(i);
 
         p.addChannel(c);
@@ -151,8 +151,8 @@ public class ObjectFactory {
         binding.setGreen(new Integer(1));
         binding.setRed(new Integer(1));
         binding.setFamily(family);
-        binding.setInputEnd(new Float(1.0));
-        binding.setInputStart(new Float(1.0));
+        binding.setInputEnd(new Double(1.0));
+        binding.setInputStart(new Double(1.0));
         binding.setNoiseReduction(Boolean.FALSE);
 
         return binding;
